@@ -3,40 +3,38 @@
   "version": "14.0.0",
   "description": "Classroom attendance management.",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "compile": "webpack --mode=production",
-    "watch": "webpack serve --mode=development"
+    "compile": "vite build",
+    "watch": "vite"
   },
   "dependencies": {
-    "cal-heatmap": "^3.6.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.4",
+    "@mui/material": "^7.3.5",
+    "cal-heatmap": "^4.2.4",
     "dayjs": "^1.11.11",
     "flux": "2.1.1",
-    "griddle-react": "0.8.x",
-    "jquery": "^3",
-    "react": "16.x",
-    "react-dom": "16.x",
-    "react-notification-system": "0.2.x",
-    "react-router": "3.x",
-    "react-skylight": "0.5.1"
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-router": "^7.9.4",
+    "react-router-dom": "^7.9.4"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.26.7",
-    "@babel/preset-react": "^7.26.3",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "babel-core": "6.x",
-    "babel-loader": "^9.2.1",
-    "babel-preset-es2015": "6.x",
-    "babel-preset-react": "6.x",
-    "babel-preset-stage-1": "latest",
-    "css-loader": "^7.1.2",
-    "eslint-webpack-plugin": "^4.2.0",
-    "html-webpack-plugin": "^5.6.3",
+    "@eslint/js": "^9.37.0",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.2",
+    "@vitejs/plugin-react": "^5.1.1",
+    "eslint": "^9.37.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.0",
     "prettier": "^3.2.5",
-    "react-refresh": "^0.14.2",
-    "style-loader": "^4.0.0",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "rollup-plugin-visualizer": "^6.0.5",
+    "vite": "^7.1.11"
+  },
+  "engines": {
+    "node": ">=22.0.0",
+    "npm": ">=11.6.0"
   }
 }
