@@ -1,10 +1,10 @@
-module.exports = {
+export default {
   printWidth: 100,
   tabWidth: 2,
   trailingComma: "all",
   singleQuote: false,
   semi: true,
-  plugins: [require.resolve("@trivago/prettier-plugin-sort-imports")],
+  plugins: ["@trivago/prettier-plugin-sort-imports"],
   importOrder: ["^[./]"],
   importOrderSeparation: true,
   importOrderSortSpecifiers: true,
