@@ -1,7 +1,6 @@
-const constants = require("./appconstants");
-const ajax = require("./ajaxhelper");
-const myhistory = require("./myhistory.js").default;
-const dispatcher = require("./appdispatcher");
+import ajax from "./ajaxhelper.js";
+import constants from "./appconstants.js";
+import dispatcher from "./appdispatcher.js";
 
 export const loadStudents = () => {
   ajax
@@ -36,7 +35,7 @@ export const updateStudent = function (id, start_date, minutes) {
     .then(loadStudentData);
 };
 
-export const swipeStudent = function (student, direction, overrideTime) {
+export const swipeStudent = function (student, direction, overrideTime, navigate) {
   let overrideDate;
   if (overrideTime) {
     overrideDate =
@@ -60,7 +59,7 @@ export const swipeStudent = function (student, direction, overrideTime) {
           message: student.name + " swiped successfully!",
         });
       }
-      myhistory.replace("/students");
+      navigate?.("/students", { replace: true });
     });
 };
 export const markAbsent = function (student) {
